Clean up authorize use case / controller action
I'm pretty satisfied with this design, where there is a use case object
for redirecting to the authorization URL and a very direct controller
action.

The integration with Settings is still a little kludgy. There should
really be a singleton config object created by an initializer or the
environment flie, rather than multiple places poking into Settings.