Move Authorizer config default from callsite to definition
The point of Github.oauth_config is that it can be expected to be there
without binding to how the values are located. Setting this as the
default in the Authorizer constructor is totally reasonable, allowing
for a different config to be injected, but not requiring other parts of
the code to supply one on all instantiations.